requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
requests==2.26.0
python-dotenv==0.19.0
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
gunicorn==20.1.0
